+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - Fixed: Task `:test` uses this output of task `:prepareSandbox` without declaring an explicit or implicit dependency [#1609](../../issues/1609)
 - ExtractorTransformer: Exclude only `Applications` symlink
 - SandboxAware: inherit sandbox directory from producer
+- Add IntelliJ Platform path-based hash to Ivy files to better deal with cache (temporary workaround)
 
 ## [2.0.0-beta3] - 2024-05-18